+++ app/src/main/java/com/example/vaccinationmanagement/doctors/Doctors.kt
@@ -1,7 +1,7 @@
 package com.example.vaccinationmanagement.doctors
 
 data class Doctors(
-    var id: Long,
+    var id: Int,
     var name: String,
     var surname: String,
 )
