+++ app/src/main/java/com/example/vaccinationmanagement/doctors/DoctorsDAO.kt
@@ -1,13 +1,13 @@
 package com.example.vaccinationmanagement.doctors
 
 interface DoctorsDAO {
-    fun getDoctorById(id: Long): Doctors?
+    fun getDoctorById(id: Int): Doctors?
 
     fun getAllDoctors(): Set<Doctors?>?
 
     fun insertDoctor(doctor: Doctors) : Boolean
 
-    fun updateDoctor(id: Long, doctor: Doctors) : Boolean
+    fun updateDoctor(id: Int, doctor: Doctors) : Boolean
 
-    fun deleteDoctor(id: Long) : Boolean
+    fun deleteDoctor(id: Int) : Boolean
 }
